added script to add data, and updated the README.md as well

+++ game_store_app/models.py
@@ -10,6 +10,6 @@ class Game(models.Model):
     game_id = models.AutoField(primary_key=True)
     title = models.CharField(max_length=255)
     platform = models.CharField(max_length=255)
-    score = models.IntegerField()
+    score = models.FloatField()
     genre = models.CharField(max_length=255)
     editors_choice = models.BooleanField()
